backend/requirements.txt
flask
flask-cors
pymongo==4.6.1
numpy==1.24.4
faiss-cpu==1.7.4
requests
flasgger==0.9.7.1
transformers==4.36.2
boto3
gunicorn==22.0.0
torch --index-url https://download.pytorch.org/whl/cpu
